src/requirements.txt
azure-mgmt-netapp>=0.5.0
azure-mgmt-resource>=3.1.0
haikunator